fixed issue with user input
before, when user was prompted to enter either w, to enter a string, or q to quit, if they entered a string of invalid characters, such as "123", the program would treat each character as a separate invalid input, and would have multiple error messages. additionally, if they entered a string that contains w, such as "w123", the program would treat the w as a valid input and try to use "123" as the input string for the computation. however, 1,2, and 3 may not be part of the input alphabet.

now, the input is treated as a string and the first character of the string is used. now, an input of "123" would result in one error message, and "w123" would be treated as "w".