progressing towards computing nfa strings
some code has been added but not all functional
most nonfunctional code copied from dfa code and is a placeholder
mostly just changed dfa to nfa
will in the future review how to make an abstract class and have two classes extend from it so i can reuse similar functions for parsing(most of the functions are very similar)
but parsing nfa text files works very similarly to dfas
has the same context free grammar
but the rules are somewhat different as far as restrictions on valid transitions (such info cant be described by the cfg)
i.e. for a dfa there have to be exactly m*n transitions, where m is the number of states and n is the size of the input alphabet set but for nfas there can be any number of transitions (including 0).